Split repo and workspace postman generation
