+++ postman/0/intercom-api.postman_collection.json
@@ -8,7 +8,7 @@
       "url": "https://developers.intercom.com"
     },
     "schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json",
-    "updatedAt": "2025-04-14T17:18:02.805Z"
+    "updatedAt": "2025-04-15T08:50:39.409Z"
   },
   "item": [
     {
@@ -2265,28 +2265,28 @@
   "event": [],
   "variable": [
     {
-      "id": "1df8eb1e-835e-4b9e-956a-7823885b9900",
+      "id": "1808d7f0-64b9-4a74-bf50-4089ce98bc4e",
       "key": "https://api.intercom.io",
       "value": "https://api.intercom.io",
       "type": "string",
       "description": "The production API server"
     },
     {
-      "id": "00a37907-c559-4bcd-b6fd-ff9fc7a69f01",
+      "id": "224773b9-223f-47d1-94a7-b7660375a996",
       "key": "https://api.eu.intercom.io",
       "value": "https://api.eu.intercom.io",
       "type": "string",
       "description": "The european API server"
     },
     {
-      "id": "47cef668-e8bc-4ef6-a3b6-fa083b06ee27",
+      "id": "1bf08464-e993-4f9f-a30e-a0f7cd85b1b6",
       "key": "https://api.au.intercom.io",
       "value": "https://api.au.intercom.io",
       "type": "string",
       "description": "The australian API server"
     },
     {
-      "id": "bbe556c6-8167-43cd-b042-833e9ad958a8",
+      "id": "cb28da62-b25d-4931-ab22-1a29bc0c1d40",
       "key": "bearerToken",
       "value": "{{bearerToken}}",
       "type": "string",
