+++ postman/2.10/intercom-api.postman_collection.json
@@ -8,7 +8,7 @@
       "url": "https://developers.intercom.com"
     },
     "schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json",
-    "updatedAt": "2025-04-14T17:18:03.037Z"
+    "updatedAt": "2025-04-15T08:50:39.646Z"
   },
   "item": [
     {
@@ -1883,28 +1883,28 @@
   "event": [],
   "variable": [
     {
-      "id": "07a19c7c-2245-447e-aeb1-b79e0836bc29",
+      "id": "2949cc3f-33c3-47e5-ac44-8be1760a262a",
       "key": "https://api.intercom.io",
       "value": "https://api.intercom.io",
       "type": "string",
       "description": "The production API server"
     },
     {
-      "id": "dc28e210-924e-411c-9679-5f2a1e3f89c7",
+      "id": "02b943c8-67cc-4d32-915f-439863c858bf",
       "key": "https://api.eu.intercom.io",
       "value": "https://api.eu.intercom.io",
       "type": "string",
       "description": "The european API server"
     },
     {
-      "id": "bb166cc6-3eb8-47c8-878c-832d61598442",
+      "id": "fe230d6c-0cc0-4358-b7b9-5886bd650e38",
       "key": "https://api.au.intercom.io",
       "value": "https://api.au.intercom.io",
       "type": "string",
       "description": "The australian API server"
     },
     {
-      "id": "ebd0468d-d937-4d2f-bded-7a04d4bd725e",
+      "id": "3c7d29b7-daa4-4708-a1bd-199f24993b2a",
       "key": "bearerToken",
       "value": "{{bearerToken}}",
       "type": "string",
