+++ postman/2.11/intercom-api.postman_collection.json
@@ -8,7 +8,7 @@
       "url": "https://developers.intercom.com"
     },
     "schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json",
-    "updatedAt": "2025-04-14T17:18:03.287Z"
+    "updatedAt": "2025-04-15T08:50:39.899Z"
   },
   "item": [
     {
@@ -1883,28 +1883,28 @@
   "event": [],
   "variable": [
     {
-      "id": "bf2ba16e-675c-4f29-8002-869d61f0c999",
+      "id": "a31ae2e5-73af-43f8-8e39-819aba298a59",
       "key": "https://api.intercom.io",
       "value": "https://api.intercom.io",
       "type": "string",
       "description": "The production API server"
     },
     {
-      "id": "f32fbeea-1d81-4eef-87a2-d50f2e131f6a",
+      "id": "94092b63-9142-4d11-8225-eb1c923a5716",
       "key": "https://api.eu.intercom.io",
       "value": "https://api.eu.intercom.io",
       "type": "string",
       "description": "The european API server"
     },
     {
-      "id": "9eb56e52-e41e-4b36-97bf-90564a6a14b3",
+      "id": "6fbcd371-c07b-41cd-8470-78747e8d212f",
       "key": "https://api.au.intercom.io",
       "value": "https://api.au.intercom.io",
       "type": "string",
       "description": "The australian API server"
     },
     {
-      "id": "ca4a2fe2-e517-4bdb-9a28-c46d9556e321",
+      "id": "f3663227-ce64-4e4a-b9fb-77ae976bc057",
       "key": "bearerToken",
       "value": "{{bearerToken}}",
       "type": "string",
