+++ postman/2.12/intercom-api.postman_collection.json
@@ -8,7 +8,7 @@
       "url": "https://developers.intercom.com"
     },
     "schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json",
-    "updatedAt": "2025-04-14T17:18:03.505Z"
+    "updatedAt": "2025-04-15T08:50:40.113Z"
   },
   "item": [
     {
@@ -2057,28 +2057,28 @@
   "event": [],
   "variable": [
     {
-      "id": "4dcd5669-2651-485c-b851-e4f519cb4133",
+      "id": "064abe9c-d9a4-4ff9-b5f6-f43f4393fc1f",
       "key": "https://api.intercom.io",
       "value": "https://api.intercom.io",
       "type": "string",
       "description": "The production API server"
     },
     {
-      "id": "0d62199e-00d5-49e7-98cb-3f5d950eb5b3",
+      "id": "7ea2ce93-0b5f-4f96-b5da-f273b24ff6cc",
       "key": "https://api.eu.intercom.io",
       "value": "https://api.eu.intercom.io",
       "type": "string",
       "description": "The european API server"
     },
     {
-      "id": "7cde7a32-1e51-4303-9f5e-b6339b9470eb",
+      "id": "32c6dc98-068f-46b0-9779-8d3514f93c7d",
       "key": "https://api.au.intercom.io",
       "value": "https://api.au.intercom.io",
       "type": "string",
       "description": "The australian API server"
     },
     {
-      "id": "e965e7d5-e603-4afb-8e5b-12a854a09683",
+      "id": "c76fdcec-1df9-4a46-8062-b44cb4ccaf5b",
       "key": "bearerToken",
       "value": "{{bearerToken}}",
       "type": "string",
