+++ postman/2.13/intercom-api.postman_collection.json
@@ -8,7 +8,7 @@
       "url": "https://developers.intercom.com"
     },
     "schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json",
-    "updatedAt": "2025-04-14T17:18:03.728Z"
+    "updatedAt": "2025-04-15T08:50:40.344Z"
   },
   "item": [
     {
@@ -2214,28 +2214,28 @@
   "event": [],
   "variable": [
     {
-      "id": "dd100048-113f-447a-9580-3bf1d98e6bb8",
+      "id": "537ca271-db3c-4394-9b43-09a149cb4ea9",
       "key": "https://api.intercom.io",
       "value": "https://api.intercom.io",
       "type": "string",
       "description": "The production API server"
     },
     {
-      "id": "c39760a6-caad-4e7a-8151-986c987907fb",
+      "id": "8f95453f-b210-45e4-9b84-877aadf5ac72",
       "key": "https://api.eu.intercom.io",
       "value": "https://api.eu.intercom.io",
       "type": "string",
       "description": "The european API server"
     },
     {
-      "id": "4d169237-905f-4f13-ac1a-10733feb5712",
+      "id": "f6df9d03-dd5e-41ae-bfae-8b97e4b2606b",
       "key": "https://api.au.intercom.io",
       "value": "https://api.au.intercom.io",
       "type": "string",
       "description": "The australian API server"
     },
     {
-      "id": "c905ec59-b3ab-4f94-81be-317fda48b7f4",
+      "id": "1b80ea9a-4631-4b58-bbcd-997aa99d45de",
       "key": "bearerToken",
       "value": "{{bearerToken}}",
       "type": "string",
