+++ postman/2.7/intercom-api.postman_collection.json
@@ -8,7 +8,7 @@
       "url": "https://developers.intercom.com"
     },
     "schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json",
-    "updatedAt": "2025-04-14T17:18:03.897Z"
+    "updatedAt": "2025-04-15T08:50:40.507Z"
   },
   "item": [
     {
@@ -1711,28 +1711,28 @@
   "event": [],
   "variable": [
     {
-      "id": "c5a8a620-a2ee-466f-9cd1-1485a9038a8e",
+      "id": "cc895296-9d8e-491e-9c7b-99e9e0ad3126",
       "key": "https://api.intercom.io",
       "value": "https://api.intercom.io",
       "type": "string",
       "description": "The production API server"
     },
     {
-      "id": "5fb96d4a-9cc6-4130-a8d0-c786faeed303",
+      "id": "b63d252a-1d3d-4d4c-b8b1-1c024b94fddd",
       "key": "https://api.eu.intercom.io",
       "value": "https://api.eu.intercom.io",
       "type": "string",
       "description": "The european API server"
     },
     {
-      "id": "e0577c61-8318-4efd-a669-54ee2bd4d89d",
+      "id": "91ac72b2-805b-48db-b6c2-16ba249cf59c",
       "key": "https://api.au.intercom.io",
       "value": "https://api.au.intercom.io",
       "type": "string",
       "description": "The australian API server"
     },
     {
-      "id": "92f6e4d0-1266-40a5-aa75-285904873880",
+      "id": "74117ede-2d23-400b-8af4-2ac57dcf8fb2",
       "key": "bearerToken",
       "value": "{{bearerToken}}",
       "type": "string",
