+++ postman/2.8/README.md
@@ -12,4 +12,4 @@ This directory contains the Postman collection for Intercom API version 2.8.
 3. Set your access token in the environment variables
 4. Start making API calls!
 
-Last updated: 2025-04-14T17:18:04.140Z
+Last updated: 2025-04-15T08:50:40.691Z
