+++ postman/2.8/intercom-api.postman_collection.json
@@ -8,7 +8,7 @@
       "url": "https://developers.intercom.com"
     },
     "schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json",
-    "updatedAt": "2025-04-14T17:18:04.138Z"
+    "updatedAt": "2025-04-15T08:50:40.689Z"
   },
   "item": [
     {
@@ -1711,28 +1711,28 @@
   "event": [],
   "variable": [
     {
-      "id": "b863b1e6-7793-4476-9799-f31efcd0b6b7",
+      "id": "cc6c36ce-d845-4444-bfc2-6aef5dd554ed",
       "key": "https://api.intercom.io",
       "value": "https://api.intercom.io",
       "type": "string",
       "description": "The production API server"
     },
     {
-      "id": "eac8fe9a-dc74-4fc5-8e02-94b7c49be2b9",
+      "id": "19d56820-d485-4da0-9ef4-e918091ee18e",
       "key": "https://api.eu.intercom.io",
       "value": "https://api.eu.intercom.io",
       "type": "string",
       "description": "The european API server"
     },
     {
-      "id": "ef24512b-3b88-4661-ba3e-6601878b1976",
+      "id": "9e2794a9-b983-4fc1-a802-19fb162a0a23",
       "key": "https://api.au.intercom.io",
       "value": "https://api.au.intercom.io",
       "type": "string",
       "description": "The australian API server"
     },
     {
-      "id": "22c0c711-e024-4bac-a315-5d83cc9299f7",
+      "id": "5d040639-9b1e-4fae-aa84-bf68084cc70c",
       "key": "bearerToken",
       "value": "{{bearerToken}}",
       "type": "string",
