+++ postman/2.9/intercom-api.postman_collection.json
@@ -8,7 +8,7 @@
       "url": "https://developers.intercom.com"
     },
     "schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json",
-    "updatedAt": "2025-04-14T17:18:04.318Z"
+    "updatedAt": "2025-04-15T08:50:40.869Z"
   },
   "item": [
     {
@@ -1882,28 +1882,28 @@
   "event": [],
   "variable": [
     {
-      "id": "4e929133-82e4-4956-8d88-2d725958cdfc",
+      "id": "bc128d90-5301-47a4-8778-52ff2524d9d9",
       "key": "https://api.intercom.io",
       "value": "https://api.intercom.io",
       "type": "string",
       "description": "The production API server"
     },
     {
-      "id": "a76dbae7-8235-4876-98fc-3785b580f5a2",
+      "id": "821fa415-e44b-4341-8361-e63ddca475bb",
       "key": "https://api.eu.intercom.io",
       "value": "https://api.eu.intercom.io",
       "type": "string",
       "description": "The european API server"
     },
     {
-      "id": "e217b43b-9a89-4756-a604-5cab470a7ed3",
+      "id": "3053b1c2-d745-4cf2-b469-943f17af24e6",
       "key": "https://api.au.intercom.io",
       "value": "https://api.au.intercom.io",
       "type": "string",
       "description": "The australian API server"
     },
     {
-      "id": "68033009-d9d1-4a82-8acd-f995650ce9e9",
+      "id": "22147da2-e3ef-4fa0-ba0e-08a5cf7b6483",
       "key": "bearerToken",
       "value": "{{bearerToken}}",
       "type": "string",
