+++ scripts/postman/run-postman-collection.js
@@ -1,37 +1,26 @@
+const uploadCollectionsToPostman = require('./upload-collections-to-postman');
 const createCollectionFromYaml = require('./create-collection-from-yaml');
+const core = require('@actions/core');
 const fs = require('fs');
-const path = require('path');
+const OUTPUT_FOLDER = './compiled';
 
-const processAllVersions = async () => {
+(async () => {
   try {
-    const descriptionsPath = path.join(process.cwd(), 'descriptions');
-
-    // Get only numeric version directories
-    const versionDirs = fs
-      .readdirSync(descriptionsPath)
-      .filter((item) => {
-        const fullPath = path.join(descriptionsPath, item);
-        return fs.statSync(fullPath).isDirectory() && !isNaN(parseFloat(item));
-      })
-      .sort((a, b) => parseFloat(a) - parseFloat(b));
-
-    console.log(`Processing versions: ${versionDirs.join(', ')}`);
-
-    for (const version of versionDirs) {
-      const yamlPath = path.join(descriptionsPath, version, 'api.intercom.io.yaml');
-      if (fs.existsSync(yamlPath)) {
-        try {
-          await createCollectionFromYaml(yamlPath);
-          console.log(`✓ Processed version ${version}`);
-        } catch (err) {
-          console.error(`Error processing version ${version}:`, err);
-        }
+    let deployCollection = false;
+    const files = process.argv.slice(2);
+    for (const file of files) {
+      if (file.slice(-5) == '.yaml') {
+        await createCollectionFromYaml(file);
+        deployCollection = true;
       }
     }
+    if (deployCollection && fs.existsSync(OUTPUT_FOLDER)) {
+      await uploadCollectionsToPostman();
+    } else {
+      console.log('No yaml files changed. Skipping collection generation.');
+    }
   } catch (err) {
-    console.error('Failed to process versions:', err);
-    process.exit(1);
+    core.setFailed('Action Failed. See logs for error.');
+    console.error(err);
   }
-};
-
-processAllVersions();
+})();
